requirements.txt
streamlit==1.32.0
agno==0.1.0
python-dotenv==1.0.1 